requirements.txt
streamlit
numpy
faiss-cpu
sentence-transformers
langchain
langchain-community
langchain-core
langchain-text-splitters
pypdf
google-generativeai
